pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "market-values"
version = "0.1.0"
description = "Live Indian stock market tracker with a web dashboard"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Jishnu T S", email = "jishnusaseendran@gmail.com" },
]
keywords = ["stock-market", "india", "nse", "dashboard", "yfinance", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "yfinance>=0.2.0",
    "pandas>=1.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
market-values = "market_values.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/market_values"]
